Split UI benchmarking tests into two files
